requirements.txt
rembg
Pillow
vtracer
pyinstaller
